SimTower clone using HTML5, CSS, and JavaScript
This is based on a project my friends and I worked on since 2008.
OpenTower was a C++ based game that used SFML. Some people decided to
revive it and made Highrise Developer, a C++ based game that used
OpenGL. Both of those projects have been dead for years so I decided to
start back up with a web friendly version.